rewrote editor
