+++ useful/index.html
@@ -5,10 +5,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Level Editor</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.9.4/p5.min.js"></script>
+    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.9.4/p5.min.js"></script> -->
     <link rel="stylesheet" href="style.css">
 
-    <script src="/collisions.js"></script>
+    <!-- <script src="/collisions.js"></script>
     <script src="/camera.js"></script>
 
     <script src="/objects/hole.js"></script>
@@ -26,12 +26,13 @@
 
     <script src="./ui/knot.js"></script>
     <script src="./ui/single.js"></script>
-    <script src="./ui/polygon.js"></script>
+    <script src="./ui/polygon.js"></script> -->
 </head>
 
 <body>
     <div class="canvas-content"></div>
-    <script src="editor.js"></script>
+    <script src="./editor.js" type="module"></script>
+    <!-- <script src="editor.js"></script> -->
 </body>
 
 </html>
