+++ useful/ui/single.js
@@ -1,7 +1,9 @@
+import { Knot } from "./knot";
+
 /**
  * Components that require only a single knot
  */
-class Single {
+export class Single {
     constructor(obj, isStatic = false) {
         this.obj = obj;
         this.knot = new Knot(obj.pos.x, obj.pos.y, this);
